+++ lib/installer_templates/example_app_tests/test/models/example_app_form_element_showcase_test.rb
@@ -36,7 +36,8 @@ class ExampleAppFormElementShowcaseTest < ActiveSupport::TestCase
     cover
     header_rich
     description
-    roles
+    locales
+    locales_display
     header_meta
     created_at
     updated_at